+++ app/[locale]/template/[templateId]/page.tsx
@@ -100,7 +100,7 @@ const TemplateEditPage: React.FC = () => {
     {
       variables: { templateId: Number(templateId) },
       notifyOnNetworkStatusChange: true,
-      fetchPolicy: "network-only",// Fetch latest data instead of cache so we get the latest sections and questions
+      fetchPolicy: "network-only",// Fetch latest data instead of cache so we get the latest sections and questions after adding new ones
     }
   );