+++ CHANGELOG.md
@@ -14,6 +14,11 @@
 ### Added
 - Added new `SectionEditContainer` component, for use in `QuestionEditCard`, to make reordering of questions more smooth [#207]
 - Added new `updateQuestionDisplayorder` and `updateSectionDisplayOrder` server actions to update question and section order when users click the `onMoveUp` and `onMoveDown` arrows [#207]
+- Added question type components to the components/Form/QuestionComponents directory [#157]
+  - Updated the `QuestionAdd` and `QuestionEdit` components to allow for the adding and editing of the new question types [#157]
+  - Updated `QuestionView` component, which is used to show a Preview of the Question that is being added or edited. Added handlers to allows users to interact with the questions in the Preview mode.
+  - Built a `MultiSelect` component using an accessible approach [#157]
+  - Created the `Number` and `Currency` components using React Aria Component's `NumberField`. This allows users to increment and decrement the field using buttons.[#157]
 - Added support for question type JSON objects [#535]
 - Added TinyMCE Editor component and added it to the styleguide [#462]
 - Added a Load more button to the Add new section page [#450]